+++ XmlParser.php
@@ -22,7 +22,16 @@ class XmlParser extends Object implements ParserInterface
      */
     public function parse(Response $response)
     {
-        return $this->convertXmlToArray($response->getContent());
+        $contentType = $response->getHeaders()->get('content-type', '');
+        if (preg_match('/charset=(.*)/i', $contentType, $matches)) {
+            $encoding = $matches[1];
+        } else {
+            $encoding = 'UTF-8';
+        }
+
+        $dom = new \DOMDocument('1.0', $encoding);
+        $dom->loadXML($response->getContent(), LIBXML_NOCDATA);
+        return $this->convertXmlToArray(simplexml_import_dom($dom->documentElement));
     }
 
     /**