+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 HTTP client extension Change Log
 2.0.4 under development
 -----------------------
 
-- no changes in this release.
+- Bug #94: Fixed `XmlParser` does not respects character encoding from response headers (klimov-paul)
 
 
 2.0.3 February 15, 2017